# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Bookworm API
  version: 0.1.0
  description: Mandatory paths/schemas per MASTER-BUILD-SPEC section 9. All endpoints versioned under /v1.
servers:
  - url: http://localhost:3001
...
paths:
...
  /v1/ai/jobs:
    get:
      tags: [v1]
      operationId: listAiJobs
      summary: List saved text AI reviews for a book
      description: Returns a bounded, tenant-scoped review history. Raw prompts, idempotency keys, provider input and private retrieval contents are never returned.
      parameters:
        - name: bookId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 20, default: 8 }
      responses:
        "200":
          description: Saved AI review summaries
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    post:
      tags: [v1]
      operationId: createAiJob
      summary: Create AI job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAiJobRequest" }
      responses:
        "201":
          description: Created AI job
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/ai/jobs/{jobId}:
    get:
      tags: [v1]
      operationId: getAiJob
      summary: Get AI job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: AI job
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/ai/suggestions/{id}/apply:
    post:
      tags: [v1]
      operationId: applyAiSuggestion
      summary: Apply suggestion
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Suggestion applied
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateAiJobRequest:
      type: object
      required: [bookId, chapterIds, agentType, idempotencyKey]
      properties:
        bookId: { type: string, format: uuid }
        chapterIds:
          type: array
          minItems: 1
          maxItems: 5
          uniqueItems: true
          items: { type: string, format: uuid }
        agentType: { type: string, enum: [writer, proofreader, copyeditor, consistency] }
        userInstruction: { type: string, minLength: 1, maxLength: 4000 }
        idempotencyKey: { type: string, minLength: 8, maxLength: 200 }
        contextPolicy:
          type: object
          additionalProperties: false
          properties:
            includeBookBible: { type: boolean, default: true }
            includeStyleGuide: { type: boolean, default: true }
            includeRelatedContext:
              type: boolean
              default: true
              description: Retrieve cited, current book sources for writer and consistency jobs.
            semanticTopK: { type: integer, minimum: 1, maximum: 20, default: 5 }
            maxTokens: { type: integer, minimum: 256, maximum: 16000, default: 4096 }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }
...
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
security:
  - bearerAuth: []